Center Map on current location when requested
